update README.md and move go files into openssl subdirectory
