+++ openssl/chacha20poly1305_test.go
@@ -16,7 +16,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/microsoft/go-crypto-openssl"
+	"github.com/microsoft/go-crypto-openssl/openssl"
 )
 
 const (
