+++ openssl/hash.go
