+++ openssl/hmac.go
