+++ openssl/pbkdf2_test.go
@@ -8,7 +8,7 @@ import (
 	"hash"
 	"testing"
 
-	"github.com/microsoft/go-crypto-openssl"
+	"github.com/microsoft/go-crypto-openssl/openssl"
 )
 
 type testVector struct {
